+++ misc/make-release.py
@@ -17,6 +17,10 @@
 #     limitations under the License.
 #
 
+""" Release script for manual usage.
+
+"""
+
 import os
 import sys
 
@@ -29,7 +33,11 @@ sys.path.insert(0, os.path.normpath(os.path.join(os.path.dirname(__file__), ".."
 import shutil
 from optparse import OptionParser
 
-from nuitka.tools.release.Debian import checkChangeLog
+from nuitka.tools.release.Debian import (
+    checkChangeLog,
+    cleanupTarfileForDebian,
+    runPy2dsc,
+)
 from nuitka.tools.release.Documentation import createReleaseDocumentation
 from nuitka.tools.release.Release import (
     checkAtHome,
@@ -39,16 +47,6 @@ from nuitka.tools.release.Release import (
 from nuitka.Version import getNuitkaVersion
 
 parser = OptionParser()
-
-parser.add_option(
-    "--use-as-ds-source",
-    action="store",
-    dest="ds_source",
-    default=None,
-    help="""\
-When given, use this as the source for the Debian package instead. Default \
-%default.""",
-)
 
 options, positional_args = parser.parse_args()
 
@@ -80,7 +78,7 @@ shutil.rmtree("dist", ignore_errors=True)
 shutil.rmtree("build", ignore_errors=True)
 
 createReleaseDocumentation()
-assert 0 == os.system("python setup.py sdist --formats=bztar,gztar,zip")
+assert os.system("python setup.py sdist --formats=bztar,gztar,zip") == 0
 
 os.chdir("dist")
 
@@ -103,70 +101,20 @@ for filename in os.listdir("."):
     if filename.endswith(".tar.gz"):
         new_name = filename[:-7] + "+ds.tar.gz"
 
-        shutil.copy(filename, new_name)
-        assert 0 == os.system("gunzip " + new_name)
-        assert 0 == os.system(
-            "tar --wildcards --delete --file "
-            + new_name[:-3]
-            + " Nuitka*/*.pdf"
-            + " Nuitka*/build/inline_copy"
-        )
-        assert 0 == os.system("gzip -9 -n " + new_name[:-3])
+        cleanupTarfileForDebian(filename, new_name)
 
-        assert 0 == os.system("py2dsc " + new_name)
-
-        # Fixup for py2dsc not taking our custom suffix into account, so we need
-        # to rename it ourselves.
-        before_deb_name = filename[:-7].lower().replace("-", "_")
-        after_deb_name = before_deb_name.replace("rc", "~rc")
-
-        assert 0 == os.system(
-            "mv 'deb_dist/%s.orig.tar.gz' 'deb_dist/%s+ds.orig.tar.gz'"
-            % (before_deb_name, after_deb_name)
-        )
-
-        assert 0 == os.system("rm -f deb_dist/*_source*")
-
-        # Remove the now useless input, py2dsc has copied it, and we don't
-        # publish it.
-        os.unlink(new_name)
-
-        if options.ds_source is not None:
-            shutil.copyfile(
-                options.ds_source, "deb_dist/%s+ds.orig.tar.gz" % after_deb_name
-            )
+        entry = runPy2dsc(filename, new_name)
 
         break
 else:
     assert False
 
 os.chdir("deb_dist")
-
-# Assert that the unpacked directory is there and file it. Otherwise fail badly.
-entry = None
-for entry in os.listdir("."):
-    if (
-        os.path.isdir(entry)
-        and entry.startswith("nuitka")
-        and not entry.endswith(".orig")
-    ):
-        break
-
-if entry is None:
-    assert False
-
-# Import the "debian" directory from above. It's not in the original tar and
-# overrides or extends what py2dsc does.
-assert 0 == os.system(
-    "rsync -a --exclude pbuilder-hookdir ../../debian/ %s/debian/" % entry
-)
-
-assert 0 == os.system("rm *.dsc *.debian.tar.xz")
 os.chdir(entry)
 
 # Build the debian package, but disable the running of tests, will be done later
 # in the pbuilders.
-assert 0 == os.system("debuild --set-envvar=DEB_BUILD_OPTIONS=nocheck")
+assert os.system("debuild --set-envvar=DEB_BUILD_OPTIONS=nocheck") == 0
 
 os.chdir("../../..")
 
@@ -176,7 +124,7 @@ assert os.path.exists("dist/deb_dist")
 
 # Check with pylint in pedantic mode and don't proceed if there were any
 # warnings given. Nuitka is lintian clean and shall remain that way.
-assert 0 == os.system("lintian --pedantic --fail-on-warnings dist/deb_dist/*.changes")
+assert os.system("lintian --pedantic dist/deb_dist/*.changes") == 0
 
 os.system("cp dist/deb_dist/*.deb dist/")
 
@@ -187,9 +135,9 @@ for filename in os.listdir("dist/deb_dist"):
 # Sign the result files. The Debian binary package was copied here.
 for filename in os.listdir("dist"):
     if os.path.isfile("dist/" + filename):
-        assert 0 == os.system("chmod 644 dist/" + filename)
-        assert 0 == os.system(
-            "gpg --local-user 2912B99C --detach-sign dist/" + filename
+        assert os.system("chmod 644 dist/" + filename) == 0
+        assert (
+            os.system("gpg --local-user 2912B99C --detach-sign dist/" + filename) == 0
         )
 
 # Cleanup the build directory, not needed.