+++ nuitka/PythonVersions.py
@@ -33,7 +33,7 @@ def getSupportedPythonVersions():
 
     """
 
-    return ("2.6", "2.7", "3.3", "3.4", "3.5", "3.6", "3.7")
+    return ("2.6", "2.7", "3.3", "3.4", "3.5", "3.6", "3.7", "3.8")
 
 
 def getPartiallySupportedPythonVersions():
@@ -41,7 +41,7 @@ def getPartiallySupportedPythonVersions():
 
     """
 
-    return ("3.8",)
+    return ()
 
 
 def getSupportedPythonVersionStr():
@@ -58,7 +58,8 @@ def getSupportedPythonVersionStr():
 def _getPythonVersion():
     big, major, minor = sys.version_info[0:3]
 
-    return big * 100 + major * 10 + minor
+    # TODO: Give up on decimal versions already.
+    return big * 100 + major * 10 + min(9, minor)
 
 
 python_version = _getPythonVersion()
@@ -149,7 +150,10 @@ def isUninstalledPython():
         system_path = os.path.normcase(buf.value)
         return not getRunningPythonDLLPath().startswith(system_path)
 
-    return "Anaconda" in sys.version or "WinPython" in sys.version
+    return (
+        os.path.exists(os.path.join(sys.prefix, "conda-meta"))
+        or "WinPython" in sys.version
+    )
 
 
 def getRunningPythonDLLPath():