+++ nuitka/plugins/standard/NumpyPlugin.py
@@ -21,16 +21,17 @@ import os
 import re
 import shutil
 import sys
-from logging import info
+from logging import info, warning
 
 from nuitka import Options
 from nuitka.plugins.PluginBase import NuitkaPluginBase
-from nuitka.utils.FileOperations import makePath
+from nuitka.plugins.Plugins import active_plugin_list
+from nuitka.utils import Execution
+from nuitka.utils.FileOperations import getFileList, makePath
 from nuitka.utils.Utils import isWin32Windows
 
 # ------------------------------------------------------------------------------
 # The following code is largely inspired by PyInstaller hook_numpy.core.py
-# ------------------------------------------------------------------------------
 # START
 # ------------------------------------------------------------------------------
 
@@ -64,13 +65,13 @@ def get_scipy_core_binaries(module):
 
 
 def get_numpy_core_binaries(module):
-    """ Return any binaries in numpy/core and/or numpy/.libs, whether or not actually used by our script.
+    """ Return any binaries in numpy/core and/or numpy/.libs.
 
     Notes:
-        This covers the special cases like MKL binaries, which cannot be detected by dependency managers.
+        This covers the special cases like MKL binaries.
 
     Returns:
-        tuple of abspaths of binaries
+        tuple of abspaths of binaries.
     """
     numpy_dir = module.getCompileTimeDirectory()
     numpy_core_dir = os.path.join(numpy_dir, "core")
@@ -126,6 +127,69 @@ def get_numpy_core_binaries(module):
 # ------------------------------------------------------------------------------
 
 
+def get_matplotlibrc():
+    """Determine the filename of matplotlibrc and the default backend.
+
+    Notes:
+        There might exist a local version outside 'matplotlib/mpl-data' which
+        we then must use instead. Determine its name by aksing matplotlib.
+    """
+    cmd = """\
+from __future__ import print_function
+from matplotlib import matplotlib_fname, get_backend
+print(matplotlib_fname())
+print(get_backend())
+"""
+
+    feedback = Execution.check_output([sys.executable, "-c", cmd])
+
+    if str is not bytes:  # ensure str in Py3 and up
+        feedback = feedback.decode()
+    feedback = feedback.replace("\r", "")
+    matplotlibrc, backend = feedback.splitlines()
+    return matplotlibrc, backend
+
+
+def copy_mpl_data(module, dist_dir):
+    """Write matplotlib data files ('mpl-data')."""
+
+    data_dir = os.path.join(module.getCompileTimeDirectory(), "mpl-data")  # must exist
+    if not os.path.isdir(data_dir):
+        sys.exit("mpl-data missing: matplotlib installation is broken")
+
+    matplotlibrc, backend = get_matplotlibrc()  # get matplotlibrc, backend
+
+    prefix = os.path.join("matplotlib", "mpl-data")
+    for item in getFileList(data_dir):  # copy data files to dist folder
+        if item.endswith("matplotlibrc"):  # handle config separately
+            continue
+        idx = item.find(prefix)  # need string starting with 'matplotlib/mpl-data'
+        tar_file = os.path.join(dist_dir, item[idx:])
+        makePath(os.path.dirname(tar_file))  # create intermediate folders
+        shutil.copyfile(item, tar_file)
+
+    old_lines = open(matplotlibrc).read().splitlines()  # old config file lines
+    new_lines = ["# modified by Nuitka plugin 'numpy'"]  # new config file lines
+    found = False  # checks whether backend definition encountered
+    for line in old_lines:
+        line = line.strip()  # omit meaningless lines
+        if line.startswith("#") or line == "":
+            continue
+        new_lines.append(line)
+        if line.startswith(("backend ", "backend:")):
+            found = True  # old config file has a backend definition
+            new_lines.append("# backend definition copied from installation")
+
+    if not found:
+        # get the string from interpreted mode and insert it in matplotlibrc
+        new_lines.append("backend: %s" % backend)
+
+    matplotlibrc = os.path.join(dist_dir, prefix, "matplotlibrc")
+    outfile = open(matplotlibrc, "w")
+    outfile.write("\n".join(new_lines))
+    outfile.close()
+
+
 class NumpyPlugin(NuitkaPluginBase):
     """ This class represents the main logic of the plugin.
 
@@ -148,10 +212,20 @@ class NumpyPlugin(NuitkaPluginBase):
         self.numpy_copied = False  # indicator: numpy files copied
         self.matplotlib = self.getPluginOptionBool("matplotlib", False)
         self.scipy = self.getPluginOptionBool("scipy", False)
+        self.scipy_copied = True  # indicator: scipy files copied
         if self.scipy:
-            self.scipy_copied = False  # indicator: scipy files copied
-        else:
-            self.scipy_copied = True
+            self.scipy_copied = False
+
+        self.mpl_data_copied = True  # indicator: matplotlib data copied
+        if self.matplotlib:
+            self.mpl_data_copied = False
+            for p in active_plugin_list:
+                if p.plugin_name.endswith("hinted-mods.py"):
+                    break
+            else:
+                warning(
+                    "matplotlib may need hinted compilation for non-standard backends"
+                )
 
     def considerExtraDlls(self, dist_dir, module):
         """ Copy extra shared libraries or data for this installation.
@@ -163,8 +237,9 @@ class NumpyPlugin(NuitkaPluginBase):
             empty tuple
         """
         full_name = module.getFullName()
+        elements = full_name.split(".")
 
-        if full_name == "numpy" and not self.numpy_copied:
+        if not self.numpy_copied and full_name == "numpy":
             self.numpy_copied = True
             binaries = get_numpy_core_binaries(module)
 
@@ -185,7 +260,7 @@ class NumpyPlugin(NuitkaPluginBase):
                 )
                 info(msg)
 
-        if full_name == "scipy" and not self.scipy_copied:
+        if not self.scipy_copied and full_name == "scipy":
             self.scipy_copied = True
             binaries = get_scipy_core_binaries(module)
 
@@ -206,84 +281,84 @@ class NumpyPlugin(NuitkaPluginBase):
                 )
                 info(msg)
 
+        if not self.mpl_data_copied and "matplotlib" in elements:
+            self.mpl_data_copied = True
+            copy_mpl_data(module, dist_dir)
+            info("Copied 'matplotlib/mpl-data'.")
+
         return ()
 
     def onModuleEncounter(self, module_filename, module_name, module_kind):
         # pylint: disable=too-many-branches,too-many-return-statements
         elements = module_name.split(".")
-        if not self.scipy and elements[0] in ("scipy", "sklearn"):
+        if not self.scipy and elements[0] in ("scipy", "sklearn", "skimage"):
+            return False, "Omit unneeded components"
+
+        if not self.matplotlib and elements[0] in ("matplotlib", "skimage"):
             return False, "Omit unneeded components"
 
         if module_name == "scipy.sparse.csgraph._validation":
             return True, "Replicate implicit import"
 
-        if elements[0] == "mpl_toolkits" and self.matplotlib is True:
+        if self.matplotlib and elements[0] == "mpl_toolkits":
             return True, "Needed by matplotlib"
-
-        if module_name.getPackageName() is None:
-            return None
 
         if module_name in ("cv2", "cv2.cv2", "cv2.data"):
             return True, "Needed for OpenCV"
 
-        if module_name in (
+        sklearn_mods = [
             "sklearn.utils.sparsetools._graph_validation",
             "sklearn.utils.sparsetools._graph_tools",
-        ):
-            return True, "Needed by sklearn"
-
-        if module_name in (
             "sklearn.utils.lgamma",
             "sklearn.utils.weight_vector",
             "sklearn.utils._unittest_backport",
-        ):
-            return True, "Needed by sklearn"
-
-        posix = (
+            "sklearn.externals.joblib.externals.cloudpickle.dumps",
             "sklearn.externals.joblib.externals.loky.backend.managers",
-            "sklearn.externals.joblib.externals.loky.backend.synchronize",
-            "sklearn.externals.joblib.externals.loky.backend.compat_posix",
-            "sklearn.externals.joblib.externals.loky.backend._posix_reduction",
-            "sklearn.externals.joblib.externals.loky.backend.popen_loky_posix",
-        )
-        win32 = (
-            "sklearn.externals.joblib.externals.loky.backend.managers",
-            "sklearn.externals.joblib.externals.loky.backend.synchronize",
-            "sklearn.externals.joblib.externals.loky.backend._win_wait",
-            "sklearn.externals.joblib.externals.loky.backend._win_reduction",
-            "sklearn.externals.joblib.externals.loky.backend.popen_loky_win32",
-        )
+        ]
 
         if isWin32Windows():
-            valid_list = win32
+            sklearn_mods.extend(
+                [
+                    "sklearn.externals.joblib.externals.loky.backend.synchronize",
+                    "sklearn.externals.joblib.externals.loky.backend._win_wait",
+                    "sklearn.externals.joblib.externals.loky.backend._win_reduction",
+                    "sklearn.externals.joblib.externals.loky.backend.popen_loky_win32",
+                ]
+            )
         else:
-            valid_list = posix
+            sklearn_mods.extend(
+                [
+                    "sklearn.externals.joblib.externals.loky.backend.synchronize",
+                    "sklearn.externals.joblib.externals.loky.backend.compat_posix",
+                    "sklearn.externals.joblib.externals.loky.backend._posix_reduction",
+                    "sklearn.externals.joblib.externals.loky.backend.popen_loky_posix",
+                ]
+            )
 
-        if module_name in valid_list:
-            return True, "Needed by sklearn"
-
-        if module_name == "sklearn.externals.joblib.externals.cloudpickle.dumps":
+        if self.scipy and module_name in sklearn_mods:
             return True, "Needed by sklearn"
 
         # some special handling for matplotlib:
-        # keep certain modules depending on whether Tk or Qt plugins are enabled
+        # depending on whether 'tk-inter' resp. 'qt-plugins' are enabled,
+        # matplotlib backends are included.
         if self.enabled_plugins is None:
             self.enabled_plugins = Options.getPluginsEnabled()
 
-        if "tk-inter" in self.enabled_plugins:
-            if module_name in (
-                "matplotlib.backends.backend_tk",
-                "matplotlib.backends.backend_tkagg",
-                "matplotlib.backend.tkagg",
-            ):
-                return True, "Needed for tkinter interaction"
+        if self.matplotlib:
+            if "tk-inter" in self.enabled_plugins:
+                if module_name in (
+                    "matplotlib.backends.backend_tk",
+                    "matplotlib.backends.backend_tkagg",
+                    "matplotlib.backend.tkagg",
+                ):
+                    return True, "Needed for tkinter backend"
 
-        if "qt-plugins" in self.enabled_plugins:
-            if module_name == "matplotlib.backends.backend_qt":
-                return True, "Needed for Qt interaction"
+            if "qt-plugins" in self.enabled_plugins:
+                if module_name.startswith("matplotlib.backends.backend_qt"):
+                    return True, "Needed for Qt backend"
 
-        if module_name == "matplotlib.backends.backend_agg":
-            return True, "Needed as standard backend"
+            if module_name == "matplotlib.backends.backend_agg":
+                return True, "Needed as standard backend"
 
 
 class NumpyPluginDetector(NuitkaPluginBase):
@@ -297,7 +372,7 @@ class NumpyPluginDetector(NuitkaPluginBase):
 
     @staticmethod
     def isRelevant():
-        """ This method is called one time only to check, whether the plugin might make sense at all.
+        """ Check whether plugin might be required.
 
         Returns:
             True if this is a standalone compilation.
@@ -305,15 +380,22 @@ class NumpyPluginDetector(NuitkaPluginBase):
         return Options.isStandaloneMode()
 
     def onModuleDiscovered(self, module):
-        """ This method checks whether a numpy module is imported.
+        """ This method checks whether numpy is required.
 
         Notes:
-            For this we check whether its full name contains the string "numpy".
+            For this we check whether its first name part is numpy relevant.
         Args:
             module: the module object
         Returns:
             None
         """
-        full_name = module.getFullName().split(".")
-        if "numpy" in full_name:
+        elements = module.getFullName().split(".")
+        if elements[0] in (
+            "numpy",
+            "scipy",
+            "skimage",
+            "pandas",
+            "matplotlib",
+            "sklearn",
+        ):
             self.warnUnusedPlugin("numpy support.")