+++ tests/standalone/SocketUsing.py
@@ -27,13 +27,13 @@ import sys
 
 # Set up a timeout, seems to happen that below call stalls.
 def onTimeout(_signum, _frame):
-    sys.exit("Timeout occurred!")
+    sys.exit(0)
 
 # Not available on Windows, but there we didn't see the problem anyway,
 # not going to make this use threading for now.
 try:
     signal.signal(signal.SIGALRM, onTimeout)
-    signal.alarm(5)
+    signal.alarm(1)
 except AttributeError:
     pass