+++ Changelog.rst
@@ -1,3 +1,17 @@
+Nuitka Release 0.6.7 (Draft)
+============================
+
+Organisational
+--------------
+
+- Added repository for Fedora 30 for download.
+
+- Updated inline copy of Scons used for Python3 to 3.1.2, which is said to
+  be faster for large compilations.
+
+The release is not done yet.
+
+
 Nuitka Release 0.6.6
 ====================
 
@@ -830,7 +844,7 @@ New Features
 Optimization
 ------------
 
-- Experimental code  for variant types for ``int`` and ``long`` values,
+- Experimental code for variant types for ``int`` and ``long`` values,
   that can be plain C value, as well as the ``PyObject *``. This is not
   yet completed though.
 
@@ -975,7 +989,7 @@ Bug Fixes
 - Standalone: Added missing implicit dependency for ``zmq`` module.
 
 - Python3.7: Fix, using the ``-X utf8`` flag on the calling interpreter, aka
-  ``--python-flag=utf8_mode`` was not  preserved in the compiled binary in all
+  ``--python-flag=utf8_mode`` was not preserved in the compiled binary in all
   cases.
 
 New Optimization
@@ -3711,7 +3725,7 @@ Cleanups
 - Connected variables with their global variable trace statically avoid the
   need to check in variable registry for it.
 
-- Removed old and mostly unused  "assume unclear locals" indications, we use
+- Removed old and mostly unused "assume unclear locals" indications, we use
   global variable traces for this now.
 
 Summary
@@ -5732,7 +5746,7 @@ Bug Fixes
   are files in standard library that are not encoded like that.
 
 - The fiber implementation for Linux amd64 was not working with glibc from
-  RHEL 5. Fixed to use  now multiple ``int`` to pass pointers as necessary.
+  RHEL 5. Fixed to use now multiple ``int`` to pass pointers as necessary.
   Also use ``uintptr_t`` instead of ``intprt_t`` to transport pointers, which
   may be more optimal.